move all options into init.lua
